change for big integer

+++ src/migrations/2500_03_27_142907_create_role_user_tables.php
@@ -15,8 +15,8 @@ class CreateRoleUserTables extends Migration
     {
         Schema::create('role_user', function (Blueprint $table) {
             $table->increments('id');
-            $table->integer('role_id')->unsigned();
-            $table->integer('user_id')->unsigned();
+            $table->bigInteger('role_id')->unsigned();
+            $table->bigInteger('user_id')->unsigned();
             $table->timestamps();
 
             $table->foreign('role_id')
